Create Isolating The Effect of One Law
function_with_graph focuses on a given law, setting it to one and receiving the mean predictions of age-adjusted rates for a number of models, then setting the law to zero and receiving the mean predictions of age-adjusted rates for the same number of models. The mean of the means is taken when the law is zero and one, and a two sample t-test is run to test if there is a significant difference between the means.
function is the same as function_with_graph, except that function_with_graph also graphs a distribution of onemeans and zeromeans.
metafunction iterates function through several model architectures.